Major refactor of all files. Cleaned up excess variables and created a service to fetch data from the database.

+++ DogDatabase/Dog.cs
@@ -65,7 +65,7 @@ namespace DogDatabase
 
         public Decimal GetRating()
         {
-            Stars = (Decimal)(Score / Ratings);
+            Stars = (Decimal)Score / (Decimal)Ratings;
             return Stars;
         }
     }
